requirements.txt
alembic==1.17.2
blinker==1.9.0
click==8.3.1
dnspython==2.8.0
email-validator==2.3.0
Flask==3.1.2
Flask-Login==0.6.3
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==3.0.3
pillow==12.1.0
psycopg2-binary==2.9.11
python-dotenv==1.2.1
qrcode==8.2
setuptools==80.9.0
SQLAlchemy==2.0.45
typing_extensions==4.15.0
Werkzeug==3.1.4
wheel==0.45.1
WTForms==3.2.1
uvicorn==0.22.0
asgiref==3.9.0
